scripts/requirements.txt
boto3==1.34.34
requests==2.31.0
Pillow==10.2.0
